feat(platform): surface active composer capabilities on chat toolbar
Enabled composer capabilities (e.g. Web search) were only visible as a ✓
inside the + menu — after closing the menu there was no signal, so users
easily forgot a capability was on. Render a pill for each active +
installed capability to the right of the model selector (both normal and
Arena mode); hover reveals an × to toggle it off.

Extracted the capability list logic from composer-mode-menu into a
shared useComposerCapabilities hook so the menu and the new pill row
stay in sync.
